+++ src/games/ss.rs
@@ -1,0 +1,7 @@
+use crate::GDResult;
+use crate::protocols::gamespy;
+use crate::protocols::gamespy::Response;
+
+pub fn query(address: &str, port: Option<u16>) -> GDResult<Response> {
+    gamespy::one::query(address, port.unwrap_or(25601), None)
+}